move crops merge into the input
