+++ src/packages/media/media/property-editors/image-cropper/property-editor-ui-image-cropper.element.ts
@@ -1,5 +1,5 @@
-import type { UmbImageCropperPropertyEditorValue } from '../../components/index.js';
-import { html, customElement, property, nothing, state, PropertyValueMap } from '@umbraco-cms/backoffice/external/lit';
+import type { UmbImageCropperPropertyEditorValue, UmbInputImageCropperElement } from '../../components/index.js';
+import { html, customElement, property, nothing, state } from '@umbraco-cms/backoffice/external/lit';
 import type { UmbPropertyEditorUiElement } from '@umbraco-cms/backoffice/extension-registry';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import '../../components/input-image-cropper/input-image-cropper.element.js';
@@ -39,33 +39,20 @@ export class UmbPropertyEditorUIImageCropperElement extends UmbLitElement implem
 	@property({ attribute: false })
 	public set config(config: UmbPropertyEditorConfigCollection | undefined) {
 		this.crops = config?.getValueByAlias<UmbImageCropperPropertyEditorValue['crops']>('crops') ?? [];
-
-		// Replace crops from the value with the crops from the config while keeping the coordinates from the value if they exist.
-		const filteredCrops = this.crops.map((crop) => {
-			const cropFromValue = this.value.crops.find((valueCrop) => valueCrop.alias === crop.alias);
-			const result = {
-				...crop,
-				coordinates: cropFromValue?.coordinates ?? undefined,
-			};
-
-			return result;
-		});
-
-		this.value = {
-			...this.value,
-			crops: filteredCrops,
-		};
 	}
 
 	#onChange(e: Event) {
-		this.value = (e.target as any).value;
+		this.value = (e.target as UmbInputImageCropperElement).value;
 		this.dispatchEvent(new UmbPropertyValueChangeEvent());
 	}
 
 	render() {
 		if (!this.value) return nothing;
 
-		return html`<umb-input-image-cropper @change=${this.#onChange} .value=${this.value}></umb-input-image-cropper>`;
+		return html`<umb-input-image-cropper
+			@change=${this.#onChange}
+			.value=${this.value}
+			.crops=${this.crops}></umb-input-image-cropper>`;
 	}
 }
